[MOVE] all components to shortnote folder
From index and library folder => shortnotes folder.
