+++ frontend/src/components/shortnotes/liList.tsx
