+++ frontend/src/components/shortnotes/snList.tsx
